requirements.txt
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.41
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
MarkupSafe==3.0.2
python-dotenv==1.0.1
pandas==2.3.1
numpy==2.3.1
openpyxl==3.1.5
xlsxwriter==3.2.5
python-docx==1.2.0
sendgrid==6.11.0
requests==2.32.3
google-generativeai==0.8.3
psycopg2-binary==2.9.9
gunicorn==22.0.0
colorama==0.4.6
lxml==6.0.0
